Add files via upload
I set up a LAMP stack on a Linux system by installing Apache as the web server, MariaDB as the database, and PHP for backend scripting. I created a simple HTML and PHP form where users can enter their ID and Name. When the form is submitted, the data is stored in a MariaDB database using PDO in PHP. I accessed the form using the server's IP address in a web browser, and verified that the submitted data was saved in the database.